requirements.txt
elasticsearch==8.16.0
fastapi==0.115.5
pandas==2.2.3
psycopg2-binary==2.9.10
uvicorn==0.32.1
